10 LLM Projects/06 Financial agent ( using phi)/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
gemini
openai